+++ docassemble_webapp/docassemble/webapp/files.py
@@ -275,6 +275,7 @@ class SavedFile(object):
                     return key.generate_url(3600)
             else:
                 #logmessage("Key " + str(keyname) + " did not exist")
+                #why not serve right from uploadedpage in this case?
                 sys.stderr.write("key " + str(keyname) + " did not exist\n")
                 return('about:blank')
         else: